refactor(entity): 统一数据类型并新增DTO类
- 将Goods、Order、OrderDetail、User实体中的Byte类型字段改为Integer类型
- 新增GoodsDTO、OrderDetailDTO数据传输对象
- 修改OrderDTO结构，移除冗余字段并添加订单明细列表
- 更新UserController返回UserDTO对象
- 调整DTO类字段命名和注释规范

+++ service-provider/src/main/java/com/zjc/provider/entity/Goods.java
@@ -54,13 +54,13 @@ public class Goods implements Serializable {
      * 状态 1上架 0下架
      */
     @TableField("status")
-    private Byte status;
+    private Integer status;
 
     /**
      * 逻辑删除
      */
     @TableField("is_deleted")
-    private Byte isDeleted;
+    private Integer isDeleted;
 
     @TableField("create_time")
     private LocalDateTime createTime;
